pyproject.toml
[tool.poetry]
name = "pyticles"
version = "0.1.0"
description = ""
authors = ["Benjamin van de Vorstenbosch <bencvdv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyopengl = "^3.1.7"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"

[tool.poetry.scripts]
pyticles = "pyticles.__main__:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
